Moved Placeholder component into its own folder
